requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
pydantic-settings==2.7.0
python-dotenv==1.0.1
alembic==1.14.0
httpx==0.28.1
google-genai==1.64.0
